changed inject logic, explained #my_each's presence
Changed the logic in inject to a case. Hopefully this makes it a little
easier to read.

Explained in comment that my_each and my_each_with_index are only
present because of convenience in regards to my goal of understanding
Enumerable better by duplicating its function.

Also fixed syntax errors on some if statements.